+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 * Look for `ZSPARQLMETHOD_DEFAULT_ENDPOINT` when creating a method [moregale]
 * Remove the `map_and_execute` method because `__call__` is enough [moregale]
+* `__call__` returns a `MethodResult` object [moregale #4516]
 
 0.3 (2011-05-31)
 ----------------